storefront/requirements.txt
Django==4.0.3
django-environ==0.8.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
Pillow==9.0.1
djangorestframework==3.13.1
django-filter==21.1
django-debug-toolbar~=3.2.4